clean up and small fixes

+++ foxspace/math.hpp
@@ -1,8 +1,10 @@
+#include <cstdint>
+
 namespace Fox {
 // returns a random number between 0 and the given one
-unsigned int get_randomNumber(unsigned int);
+uint32_t get_randomNumber(uint32_t);
 // returns true if the given number is a prime
-unsigned int is_prime(unsigned int);
+uint32_t is_prime(uint32_t);
 // returns the summed prime numbers from 1 to the given one
-unsigned int prime_sum(unsigned int);
+uint64_t prime_sum(uint32_t);
 }  // namespace Fox
